requirements.txt
pyspark~=3.4.1
py4j~=0.10.9.7
pyyaml~=6.0.1
